requirements.txt
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-autocomplete-light==3.5.1
django-cors-headers==3.1.1
django-rest-knox==4.1.0
django-stubs==1.2.0
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
djangorestframework-xml==1.4.0
freeze==1.0.10
idna==2.9
joblib==0.14.1
mypy==0.720
mypy-extensions==0.4.3
numpy==1.18.1
Pillow==7.0.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.8
python-dotenv==0.13.0
mysqlclient==1.3.13
beautifulsoup4==4.9.0
django-tabbed-admin==1.0.4